+++ src/components/prompt-list-item.tsx
@@ -17,23 +17,23 @@ import {
   Application,
 } from "@raycast/api";
 import { runAppleScript } from "@raycast/utils";
-import { PromptProps } from "../managers/PromptManager";
-import { SpecificReplacements } from "../utils/placeholderFormatter";
+import { PromptProps } from "../managers/prompt-manager";
+import { SpecificReplacements } from "../utils/placeholder-formatter";
 import path from "path";
-import { generatePromptActions } from "./PromptActions";
-import { getPlaceholderIcons, findOptionPlaceholders } from "../utils/promptFormattingUtils";
-import { ScriptInfo } from "../utils/scriptUtils";
-import { placeholderFormatter } from "../utils/placeholderFormatter";
-import { PromptList } from "./PromptList";
-import { PromptOptionsForm } from "./PromptOptionsForm";
+import { generatePromptActions } from "./prompt-actions";
+import { getPlaceholderIcons, findOptionPlaceholders } from "../utils/prompt-formatting-utils";
+import { ScriptInfo } from "../utils/script-utils";
+import { placeholderFormatter } from "../utils/placeholder-formatter";
+import { PromptList } from "./prompt-list";
+import { PromptOptionsForm } from "./prompt-options-form";
 import {
   addTemporaryDirectory,
   removeTemporaryDirectory,
   removeAllTemporaryDirectories,
   getActiveTemporaryDirectoriesWithExpiry,
   TemporaryDirectoryWithExpiry,
-} from "../stores/TemporaryPromptDirectoryStore";
-import promptManager from "../managers/PromptManager";
+} from "../stores/temporary-directory-store";
+import promptManager from "../managers/prompt-manager";
 import fs from "fs";
 
 interface QuickGPTExtensionPreferences {